+++ src/interlude/interlude2025/performance.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct Interlude2025Performance {
-    pub pp: f64,
+    pub ratings: f64,
 }
 
 impl crate::calculator::Rating for Interlude2025Performance {}