+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod calculator;
-pub mod osu;
+pub mod clock_rate;
 pub mod interlude;
+pub mod osu;